requirements.txt
appdirs==1.4.0
apt-xapian-index==0.45
BeautifulSoup==3.2.1
beautifulsoup4==4.2.1
binwalk==2.1.1
chardet==2.0.1
click==6.6
colorama==0.3.7
command-not-found==0.3
Cython==0.20.1.post0
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
dissy==9
distorm3==3.3.0
Django==1.9.7
dnspython==1.11.1
docutils==0.11
eyeD3==0.6.18
Flask==0.11.1
hachoir-core==1.3.3
hachoir-parser==1.3.4
httplib2==0.8
idna==2.1
impacket==0.9.10
itsdangerous==0.24
Jinja2==2.7.2
knockpy==3.0
lxml==3.3.3
Magic-file-extensions==0.2
Mako==0.9.1
MarkupSafe==0.18
MAT==0.4rc0
matplotlib==1.3.1
mutagen==1.31
MySQL-python==1.2.3
nose==1.3.1
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
openpyxl==1.7.0
PAM==0.4.2
paramiko==1.10.1
pcapy==0.10.8
pdfrw==0.1
pefile==1.2.9.1
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
prettytable==0.7.2
psutil==1.2.1
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
Pygments==1.6
pygobject==3.12.0
pygubu==0.9.7.8
pyliblzma==0.5.3
PyOpenGL==3.0.2
pyOpenSSL==0.13
pyparsing==2.0.1
pyqtgraph==0.9.10
pyrit==0.4.0
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==1.0.1
pytesseract==0.1.6
python-apt===0.9.3.5ubuntu2
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
python-docx==0.8.6
python-lzo==1.8
python-socksipy==1.0
pytz==2012rc0
pyxdg==0.25
PyYAML==3.10
reportlab==3.0
requests==2.2.1
requests-file==1.4
roman==2.0.0
scapy==2.2.0
scikit-image==0.9.3
scipy==0.13.3
sessioninstaller==0.0.0
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
Sphinx==1.2.2
SQLAlchemy==0.8.4
ssh-import-id==3.21
tldextract==2.0.1
Twisted-Core==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
vboxapi==1.0
vinetto==0.6a0
virtualenv==15.0.2
volatility==2.5
wapiti==2.3.0
Werkzeug==0.11.10
wxPython==2.8.12.1
wxPython-common==2.8.12.1
yara-python==3.4.0
zenmap==7.1
zim==0.60
zope.interface==4.0.5
